Add files via upload
A set of macros that standarize the interface between link_messsages (in-prim) and listen (in-world) communication.  States are represented as integers that indicated which script should be called.  A state is invoked by the second parameter to link_message, which listens for particular state numbers and takes action.  The message argument to both listen and link_message is structured.  The bar (|) character is used to divide the message into control and data elements where the first segment is control and each additional segment is data (e.g. CONTROL|DATA1|DATA2|...). The key parameter is used for any key, such as the calling agent.

Scripts are structured to expect certain data on the stack.  This data is removed (POP'ed) or added (PUSH'ed) onto the data stack by the script.  In the abstract, each state (script) is a microservice that is promised certain inputs to product specific outputs.

A prim interacts with other objects through the SAY commands and Listen.  A prim can send a string of commands to an external object, allowing for complex interactions.

Example:
A user enters a region.  A "server" on the land sees the new arrival and their location.  The server rezzes a prim and passes information about the agent and their location.  The prim warps to the agent and follows them.  The agent interacts with the prim through the chat channel,  The agent may request a teleport, where the prim contacts the server, loads available teleport locations, and products a menu for of locations for the user.  The user selects a location and is teleported elsewhere on the sim.  (note: no physical teleporter is necessary and a user can teleport from anywhere).